+++ examples/analog/src/main.rs
@@ -7,9 +7,10 @@ use panic_halt as _;
 use cortex_m_rt::entry;
 
 use microbit::{
+    adc::{Adc, AdcConfig, Default},
     board::Board,
     display::blocking::Display,
-    hal::{prelude::*, saadc::SaadcConfig, Saadc, Timer},
+    hal::{prelude::*, Timer},
 };
 
 #[entry]
@@ -17,8 +18,8 @@ fn main() -> ! {
     if let Some(board) = Board::take() {
         let mut timer = Timer::new(board.TIMER0);
         let mut display = Display::new(board.display_pins);
-        let mut adc: Saadc = Saadc::new(board.ADC, SaadcConfig::default());
-        let mut anapin = board.edge.e00.into_floating_input(); // PAD1
+        let mut adc = Adc::new(board.ADC, AdcConfig::default_10bit());
+        let mut anapin = board.edge.e00.into_floating_input(); // PAD0
 
         let numbers = [
             [