+++ microbit-common/src/adc.rs
@@ -1,0 +1,6 @@
+//! ADC
+#[cfg(feature = "v1")]
+pub use crate::v1::adc::*;
+
+#[cfg(feature = "v2")]
+pub use crate::v2::adc::*;